==> notebooks/01_scrape_data.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))
from src.scraper import second_diff, shot_scraper, get_game_ids, shot_scraper2

import pandas as pd
from nhlpy import NHLClient
from pathlib import Path

client = NHLClient()

games_2025 = get_game_ids()
games_2024 = get_game_ids([20232024])
games_2023 = get_game_ids([20222023])

client.game_center.play_by_play(2024020705)["rosterSpots"]

game_ids=[2024020705]
test = shot_scraper(game_ids)

test = shot_scraper(game_ids)

test2 = shot_scraper2(game_ids)

test3 = shot_scraper2([2024020706])

test

test.info()

shots_2025 = shot_scraper2(games_2025)

shots_2025

shots_2024 = shot_scraper2(games_2024)

shots_2023 = shot_scraper2(games_2023)

shots_2024["shooter"].value_counts()

shots_2023.info()

shots_2024.info()

shots_2025.info()

shots_2025.isna().sum()

shots_2024.isna().sum()

shots_2023.isna().sum()

path2025 = Path('..') / 'data' / 'raw' / 'shots_2025.parquet'
path2024 = Path('..') / 'data' / 'raw' / 'shots_2024.parquet'
path2023 = Path('..') / 'data' / 'raw' / 'shots_2023.parquet'

shots_2025.to_parquet(path=path2025, engine="pyarrow")
shots_2024.to_parquet(path=path2024, engine="pyarrow")
shots_2023.to_parquet(path=path2023, engine="pyarrow")
