==> notebooks/02_clean_and_fe.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

import numpy as np
import pandas as pd
from nhlpy import NHLClient
from pathlib import Path
import pyarrow

client = NHLClient()

path2025 = Path('..') / 'data' / 'raw' / 'shots_2025.parquet'
path2024 = Path('..') / 'data' / 'raw' / 'shots_2024.parquet'
path2023 = Path('..') / 'data' / 'raw' / 'shots_2023.parquet'

shots_2025 = pd.read_parquet(path2025, engine="pyarrow")
shots_2024 = pd.read_parquet(path2024, engine="pyarrow")
shots_2023 = pd.read_parquet(path2023, engine="pyarrow")

shots_2025

shooter_ids2025 = shots_2025["shooter_id"].tolist()

test_dict = client.stats.player_career_stats(8479973)
test_dict["position"]
test_dict["shootsCatches"]
test_dict["featuredStats"]["regularSeason"]["career"]

stats2025 = []
stats_dict = {}
for id in shooter_ids2025:
    if id not in stats_dict.keys():
        stats = client.stats.player_career_stats(id)
        try:
            stats_dict[id] = {
                "position": stats["position"],
                "hand": stats["shootsCatches"],
                "pct": stats["featuredStats"]["regularSeason"]["career"]["shootingPctg"]
                }
        except:
            stats_dict[id] = {
                "position": stats["position"],
                "hand": stats["shootsCatches"],
                "pct": None
                }

    position = stats_dict[id]["position"]
    shooter_hand = stats_dict[id]["hand"]
    shooting_pct = stats_dict[id]["pct"]
    stats2025.append((position, shooter_hand, shooting_pct))

len(stats2025)

header = ["position", "shooter_hand", "shooting_pct"]
stats2025_df = pd.DataFrame(stats2025, columns=header)

goalie_ids2025 = shots_2025["goalie_id"].tolist()

goalie_stats2025 = []
goalie_stats_dict = {}
for id in goalie_ids2025:
    if id not in goalie_stats_dict.keys():
        stats = client.stats.player_career_stats(id)
        try:
            goalie_stats_dict[id] = {
                "hand": stats["shootsCatches"],
                "pct": stats["featuredStats"]["regularSeason"]["career"]["savePctg"]
                }
        except:
            goalie_stats_dict[id] = {
                "hand": stats["shootsCatches"],
                "pct": None
                }

    shooter_hand = goalie_stats_dict[id]["hand"]
    save_pct = goalie_stats_dict[id]["pct"]
    goalie_stats2025.append((shooter_hand, save_pct))

goalie_header = ["glove_hand", "save_pct"]
goalie_stats2025_df = pd.DataFrame(goalie_stats2025, columns=goalie_header)

goalie_stats2025_df

complete_df2025 = pd.concat([shots_2025, stats2025_df, goalie_stats2025_df], axis=1)
complete_df2025.head()

from src.scraper import get_skater_stats

complete_df = get_skater_stats(shots_2025)

complete_df.head()

complete_df["home_skaters"].value_counts()

pd.set_option('display.max_columns', None)

complete_df.head()

# Every shot is plotted on one half of the ice, the red line is at x = 0, the blue line is at x = 25, goal line at x = 89, all measurements are in feet.
# Use this information to create an angle to the net feature
def angle(x_coord, y_coord):
    x_centered = 89 - x_coord
    return round(np.degrees(np.arctan(y_coord/x_centered)), 2)

complete_df['angles'] = angle(complete_df['x_coord'], complete_df['y_coord'])
complete_df.head()

complete_df["shot_on_glove"] = complete_df["shooter_hand"] + complete_df["glove_hand"]

complete_df["home_skaters"] = complete_df["home_skaters"].astype(int)
complete_df["away_skaters"] = complete_df["away_skaters"].astype(int)
complete_df = complete_df[complete_df["home_skaters"] >= 3]
complete_df = complete_df[complete_df["away_skaters"] >= 3]

complete_df["home_skaters"].value_counts()

complete_df["situation"] = complete_df.apply(
    lambda row: "EV" if row["home_skaters"] == row["away_skaters"] else
    ('PP' if row['home_skaters'] > row['away_skaters'] else 'SH'),
    axis=1
    )

complete_df["situation"].value_counts()

complete_df.head()

complete_df.info()

from src.scraper import get_processed_data

test_df = get_skater_stats(shots_2025)

processed_df = get_processed_data(test_df)

processed_df

df_2024 = get_skater_stats(shots_2024)

processed_2024 = get_processed_data(df_2024)

processed_2024

df_2023 = get_skater_stats(shots_2023)

processed_2023 = get_processed_data(df_2023)

processed_2023

path2025 = Path('..') / 'data' / 'processed' / 'processed_shots_2025.parquet'
path2024 = Path('..') / 'data' / 'processed' / 'processed_shots_2024.parquet'
path2023 = Path('..') / 'data' / 'processed' / 'processed_shots_2023.parquet'

processed_df.to_parquet(path=path2025, engine="pyarrow")
processed_2024.to_parquet(path=path2024, engine="pyarrow")
processed_2023.to_parquet(path=path2023, engine="pyarrow")
