==> notebooks/03_log_regression.ipynb <==
%load_ext autoreload
%autoreload 2
%config InteractiveShell.ast_node_interactivity = "all"

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

import pandas as pd
import pyarrow.parquet as pq
from pathlib import Path

path2023 = Path('..') / 'data' / 'processed' / 'processed_shots_2023.parquet'
path2024 = Path('..') / 'data' / 'processed' / 'processed_shots_2024.parquet'
path2025 = Path('..') / 'data' / 'processed' / 'processed_shots_2025.parquet'

processed2023 = pd.read_parquet(path2023, engine='pyarrow')
processed2024 = pd.read_parquet(path2024, engine='pyarrow')
processed2025 = pd.read_parquet(path2025, engine='pyarrow')

full_df = pd.concat([processed2023, processed2024, processed2025])
full_df.head()

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, classification_report

cleaned_df = full_df.dropna()
X = cleaned_df.drop(['target', 'shot_class'], axis=1)
y = cleaned_df['target']

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, test_size=0.2, random_state=42)

from sklearn.preprocessing import OneHotEncoder, MinMaxScaler, FunctionTransformer, StandardScaler
from sklearn.pipeline import Pipeline, make_pipeline
from sklearn.compose import ColumnTransformer, make_column_selector, make_column_transformer

cat_pipeline = make_pipeline(OneHotEncoder(handle_unknown="ignore"))
num_pipeline = make_pipeline(StandardScaler(), MinMaxScaler((-1,1)))

preprocessing = ColumnTransformer([
        ("cat", cat_pipeline, make_column_selector(dtype_include=object)),
    ],
    remainder=num_pipeline)

log_reg = make_pipeline(preprocessing, LogisticRegression())
log_reg.fit(X_train, y_train)
goal_predictions = log_reg.predict(X_train)
goal_predictions[:5].round(-2)
y.iloc[:5].values

probabilities_log = log_reg.predict_proba(X_test)
pred_test = log_reg.predict(X_test)

probabilities_log

from sklearn.metrics import roc_auc_score, confusion_matrix, classification_report

auc_test = roc_auc_score(y_test, probabilities_log[:,1])
auc_train = roc_auc_score(y_train, log_reg.predict_proba(X_train)[:,1])

confusion_matrix(y_test, pred_test)
print(classification_report(y_test, pred_test))

auc_test
auc_train

from src.experiment_utils import set_mlflow_tracking, log_model_to_mlflow
from dotenv import load_dotenv
load_dotenv()

mlflow = set_mlflow_tracking()

log_model_to_mlflow(log_reg, X_test, "Logistic_regression", "AUC", score=auc_test)
