==> notebooks/05_test_xG.ipynb <==
%load_ext autoreload
%autoreload 2
%config InteractiveShell.ast_node_interactivity = "all"

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

import pandas as pd
from pathlib import Path
import joblib

path = Path('..') / 'models' / 'xgb_v1.pkl'
model = joblib.load(path)

from nhlpy import NHLClient
from src.scraper import shot_scraper2

client = NHLClient()

game_ids = [2024021171]

df = shot_scraper2(game_ids)

df.head()

df.info()

from src.scraper import get_skater_stats, get_processed_data

full_df = get_skater_stats(df)
processed_df = get_processed_data(full_df)

processed_df.head()

xG = model.predict_proba(processed_df)

df['xG'] = xG[:,1]

away_df = df[df["home"] == 0]
home_df = df[df["home"] == 1]

away_df['xG'].sum()

home_df['xG'].sum()
